ui: Add missing casts

+++ payload/game/ui/Section.cc
@@ -192,7 +192,7 @@ void Section::addPage(PageId pageId) {
         }
     }
 
-    SP_LOG("addPage %u", pageId);
+    SP_LOG("addPage %u", static_cast<u32>(pageId));
 
     REPLACED(addPage)(pageId);
 }
@@ -223,13 +223,13 @@ void Section::addActivePage(PageId pageId) {
         }
     }
 
-    SP_LOG("addActivePage %u", pageId);
+    SP_LOG("addActivePage %u", static_cast<u32>(pageId));
 
     REPLACED(addActivePage)(pageId);
 }
 
 void Section::addPages(SectionId id) {
-    SP_LOG("&7DEBUG: Constructing section %u (0x%x)", id, id);
+    SP_LOG("&7DEBUG: Constructing section %u (0x%x)", static_cast<u32>(id), static_cast<u32>(id));
 
     REPLACED(addPages)(id);
